Move DoSerialise* methods to metal_serialise.cpp
PR Feedback.
Moved from metal_types.cpp -> metal_stringise.cpp
