+++ renderdoc/driver/metal/CMakeLists.txt
@@ -38,6 +38,7 @@ set(sources
     metal_init_state.cpp
     metal_helpers_bridge.h
     metal_helpers_bridge.mm
+    metal_serialise.cpp
     metal_stringise.cpp
     official/metal-cpp.h
     official/metal-cpp.cpp)
